import-maps/acquire-import-maps-flag/dynamic-import/success.tentative.html

<!DOCTYPE html>
<html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
// https://github.com/WICG/import-maps/blob/master/spec.md#when-import-maps-can-be-encountered
const t = async_test(
  'After dynamic imports, import maps should fire error events');
const log = [];
promise_test(() => import('../../resources/slow-script.py'),
             "A dynamic import succeeds");
</script>
<script type="importmap" onerror="t.done()">
{
  "imports": {
    "../../resources/log.js?pipe=sub&name=A": "../../resources/log.js?pipe=sub&name=B"
  }
}
</script>
<script>
promise_test(() => {
  return import("../../resources/log.js?pipe=sub&name=A")
    .then(() => assert_array_equals(log, ["log:A"]))
  },
  'After <script type="module"> import maps are not effective');
</script>
